The weather project I had been working on
If this messes up the main branch/pulling data from the 495weatherapp, sorry. Feel free to delete this. I just wanted to send you what I had put together from the other video. An easy way to check it out without setting up a server is the application Brackets. It's an editor that also has live preview via google chrome. I liked how this one allows for users to input a city. There are some bugs but functional enough to get the gist.